Implement PDF generation at the abstraction level.
